Transaction and clean up

+++ domains/expense/model/expensemodel.go
@@ -5,14 +5,14 @@ import (
 )
 
 type Expense struct {
-	base.Model      `gorm:"extends"`
-	Id              int64
-	UserId 			int64
-	Nama	        string  `gorm:"varchar(100)"`
-	ReleaseDate		string  `gorm:"varchar(10)"`
-	Quantity        float32 `gorm:"double"`
-	Category		int
-	Description	    string  `gorm:"varchar(255)"`
+	base.Model  `gorm:"extends"`
+	Id          int64
+	UserId      int64
+	Nama        string  `gorm:"varchar(100)"`
+	ReleaseDate string  `gorm:"varchar(10)"`
+	Quantity    float32 `gorm:"double"`
+	Category    int64
+	Description string `gorm:"varchar(255)"`
 }
 
 func init() {
